implement conveyor belt in 3a and 3b
