--- Assets/Prefabs/ConveyorBelt.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 32b4f02b661554ede8561bd89e4cf32e
-timeCreated: 1456897047
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
