+++ Assets/Scripts/ConveyorBeltPiece.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 860ff0dacc6e04323a03c9d603f256ff
+timeCreated: 1462125358
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
