+++ Assets/Scripts/Manager/levelManager.cs
@@ -168,7 +168,7 @@ public class levelManager : MonoBehaviour
 	{
 		Vector2 direction = doll.transform.position - player.transform.position;
 
-		LayerMask layer = 1 << LayerMask.NameToLayer ("Doll") | 1 << LayerMask.NameToLayer ("ghostCollider");
+		LayerMask layer = 1 << LayerMask.NameToLayer ("Doll") | 1 << LayerMask.NameToLayer ("ghostCollider") | 1 << LayerMask.NameToLayer ("Wall");
 		RaycastHit2D[] hits = Physics2D.RaycastAll (player.transform.position, direction,
 			GetComponent<DollManager> ().radius, layer);
 		Debug.DrawRay (player.transform.position, direction, Color.green, 0.3f);
