--- Assets/Sprites/Level 3/ConveyorBelt.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a7a36e779248e41daaa36eea1ce0eafa
-timeCreated: 1461467601
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
