--- Assets/Sprites/Level 3/ConveyorBelt1.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3067e9343b4704df7a1d2fc899d5964a
-timeCreated: 1461468965
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
