--- Assets/Sprites/Level 3/ConveyorBelt2.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 619c37dfa0839442b9d1faa3972ac65e
-timeCreated: 1461469508
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
