--- Assets/Sprites/Level 3/ConveyorBelt3.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: af10a5396fca547b99851b9fa51d3fa6
-timeCreated: 1461469540
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
