--- Assets/Sprites/Level 3/ConveyorBeltAnim2.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5ba1c14de27bd4c9193801c28f565cb7
-timeCreated: 1461469566
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
