--- Assets/Sprites/Level 3/ConveyorBeltAnim3.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 56f5c009286a64f30822d4181f202d74
-timeCreated: 1461469622
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
